use suffixed literals instead of casts to construct 'GLfloat' values
The `GLfloat` type is 32-bit IEEE-754 floating-point.¹ As of C99, the `float`
type is also 32-bit IEEE-754 floating-point on all platforms with support.² In
contemporary usage, this is effectively all C99 platforms that support
floating-point arithmetic. Graphviz has been compiled in C99 mode since commit
fe3f9411d2c59b463ab1b64eecfd19f2db55d2fc. Putting those things together, we can
abbreviate the way we write `GLfloat` literals, leading to more readable code.

¹ https://www.khronos.org/opengl/wiki/OpenGL_Type
² https://en.wikipedia.org/wiki/C99#IEEE_754_floating-point_support

+++ lib/glcomp/glcompimage.c
@@ -106,7 +106,7 @@ void glCompImageDraw(void *obj)
     {
 	w = p->width;
 	h = p->height;
-	d=(GLfloat)p->common.layer* (GLfloat)GLCOMPSET_BEVEL_DIFF;
+	d=(GLfloat)p->common.layer * GLCOMPSET_BEVEL_DIFF;
 	glDisable(GL_BLEND);
 	glEnable(GL_TEXTURE_2D);
 	glTexEnvf (GL_TEXTURE_ENV, GL_TEXTURE_ENV_MODE, GL_DECAL);
